+++ backend/app/models/__init__.py
@@ -4,7 +4,7 @@ from .transacao_estoque import TransacaoEstoque
 from .ncm import NCM
 from .refresh_token import RefreshToken
 from .cliente import Cliente
-from .venda import Venda, VendaItem
+from .venda import Venda, VendaItem, VendaPagamento
 from .conta_receber import ContaReceber
 from .movimentacao_estoque import MovimentacaoEstoque
 from .nota_fiscal import NotaFiscal, NotaFiscalItem